Updated search page, can be accessed from edit menu as "Add File..."
Search page can now be accessed via the edit/manage dataset page, by selecting "Add File to an Element" in the dropdown menu. Each row of the results table can be clicked to bring up the data about that row. This functionality is also available on the standard view dataset page, because why not.